+++ Changes.md
@@ -10,6 +10,7 @@
 
 - bounce: repackaged plugin as NPM module #3341
 - connection: check remote is connected before queue #3338
+  - improve log message for queue* hooks, fixes #2998
 - transaction: init with conn.init_transaction, always pass in cfg #3315
 - check for local_mx only when default route is used #3307
 - dkim: repackaged as NPM module #3311